refactor(shell): split createPlaywrightCommand into a handler-table dispatcher
Decomposes the 3,657-line playwright-command.ts (whose createPlaywrightCommand
held a single 1,720-line 44-case switch) into a thin ~110-line dispatcher plus
focused modules under playwright/: state/flags, a11y snapshot, session logging,
teleport state machine, teleport storage, Link-header discovery, help text, and
one handler module per subcommand family. No behavior change; public exports
(createPlaywrightCommand, PLAYWRIGHT_COMMAND_NAMES, getSharedState, asWebFetch,
setPlaywrightTeleport*) are preserved via re-export.

Fixes #996
