+++ biome.json
@@ -290,6 +290,7 @@
         "packages/webapp/src/shell/supplemental-commands/open-command.ts",
         "packages/webapp/src/shell/supplemental-commands/pdftk-command.ts",
         "packages/webapp/src/shell/supplemental-commands/playwright-command.ts",
+        "packages/webapp/src/shell/supplemental-commands/playwright/**",
         "packages/webapp/src/shell/supplemental-commands/secret-command.ts",
         "packages/webapp/src/shell/supplemental-commands/upskill-command.ts",
         "packages/webapp/src/shell/supplemental-commands/websocat-command.ts",
@@ -366,6 +367,7 @@
         "packages/webapp/src/shell/supplemental-commands/open-command.ts",
         "packages/webapp/src/shell/supplemental-commands/pdftk-command.ts",
         "packages/webapp/src/shell/supplemental-commands/playwright-command.ts",
+        "packages/webapp/src/shell/supplemental-commands/playwright/**",
         "packages/webapp/src/shell/supplemental-commands/ps-command.ts",
         "packages/webapp/src/shell/supplemental-commands/python-command.ts",
         "packages/webapp/src/shell/supplemental-commands/screencapture-command.ts",
